+++ src/SprintChange.php
@@ -224,6 +224,36 @@ class SprintChange extends CommonDBRelation
         echo "</table></div>";
     }
 
+    public function post_addItem()
+    {
+        $change = new Change();
+        if ($change->getFromDB($this->fields['changes_id'])) {
+            $item = new SprintItem();
+            $item->add([
+                'plugin_sprint_sprints_id' => $this->fields['plugin_sprint_sprints_id'],
+                'name'                     => $change->fields['name'],
+                'itemtype'                 => 'Change',
+                'items_id'                 => $this->fields['changes_id'],
+                'status'                   => SprintItem::STATUS_TODO,
+                'priority'                 => (int)($change->fields['priority'] ?? 3),
+                'users_id'                 => (int)($this->fields['users_id'] ?? 0),
+            ]);
+        }
+    }
+
+    public function post_purgeItem()
+    {
+        $item = new SprintItem();
+        $items = $item->find([
+            'plugin_sprint_sprints_id' => $this->fields['plugin_sprint_sprints_id'],
+            'itemtype'                 => 'Change',
+            'items_id'                 => $this->fields['changes_id'],
+        ]);
+        foreach ($items as $row) {
+            $item->delete(['id' => $row['id']], 1);
+        }
+    }
+
     public static function cleanForItem(\CommonDBTM $item): void
     {
         $temp = new self();