+++ src/SprintTicket.php
@@ -233,6 +233,36 @@ class SprintTicket extends CommonDBRelation
         echo "</table></div>";
     }
 
+    public function post_addItem()
+    {
+        $ticket = new Ticket();
+        if ($ticket->getFromDB($this->fields['tickets_id'])) {
+            $item = new SprintItem();
+            $item->add([
+                'plugin_sprint_sprints_id' => $this->fields['plugin_sprint_sprints_id'],
+                'name'                     => $ticket->fields['name'],
+                'itemtype'                 => 'Ticket',
+                'items_id'                 => $this->fields['tickets_id'],
+                'status'                   => SprintItem::STATUS_TODO,
+                'priority'                 => (int)($ticket->fields['priority'] ?? 3),
+                'users_id'                 => (int)($this->fields['users_id'] ?? 0),
+            ]);
+        }
+    }
+
+    public function post_purgeItem()
+    {
+        $item = new SprintItem();
+        $items = $item->find([
+            'plugin_sprint_sprints_id' => $this->fields['plugin_sprint_sprints_id'],
+            'itemtype'                 => 'Ticket',
+            'items_id'                 => $this->fields['tickets_id'],
+        ]);
+        foreach ($items as $row) {
+            $item->delete(['id' => $row['id']], 1);
+        }
+    }
+
     /**
      * Clean relation when a ticket is purged
      */